move data retention implementation out of pkg/app/into a dedicated pkg/system/dataretention/ package also replace COUNT with n_live_tup
