--- pkg/app/stores_data_retention_statfs_windows.go
@@ -1,8 +1,0 @@
-//go:build windows
-
-package app
-
-// sqliteDiskCapacity is not supported on Windows; always returns ok=false.
-func sqliteDiskCapacity(_ string, _ uint64) (capacity uint64, utilization float32, ok bool) {
-	return
-}
