+++ pkg/system/dataretention/statfs.go
@@ -1,13 +1,13 @@
 //go:build !windows
 
-package app
+package dataretention
 
 import "syscall"
 
-// sqliteDiskCapacity returns the total disk capacity and utilization for the filesystem
+// diskCapacity returns the total disk capacity and utilisation for the filesystem
 // containing dataDir, expressed relative to the current DB size.
 // Returns ok=false if the stats cannot be determined.
-func sqliteDiskCapacity(dataDir string, dbUsedBytes uint64) (capacity uint64, utilization float32, ok bool) {
+func diskCapacity(dataDir string, dbUsedBytes uint64) (capacity uint64, utilization float32, ok bool) {
 	var st syscall.Statfs_t
 	if err := syscall.Statfs(dataDir, &st); err != nil || st.Bsize <= 0 || st.Blocks == 0 {
 		return
